app bar as a widget
