--- lib/ui/screens/home/section_header.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_ecom_demo/ui/theme_colors.dart';
-
-class SectionHeader extends StatelessWidget {
-  const SectionHeader({
-    Key? key,
-    required this.title,
-    this.onMorePressed,
-  }) : super(key: key);
-
-  final String title;
-  final VoidCallback? onMorePressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Text(title.toUpperCase(), style: Theme.of(context).textTheme.subtitle2),
-        const Spacer(),
-        TextButton(
-          onPressed: () {},
-          child: const Text(
-            'See More',
-            style: TextStyle(color: ThemeColors.nebula),
-          ),
-        )
-      ],
-    );
-  }
-}
